+++ protocols/P2NClient/IP2NClient.h
@@ -91,7 +91,7 @@ public:
 public:
     /** Creates a client.
     */
-    static IP2NClient* createClient(const std::string& host, int port, int verbosity, const std::string& extras);
+    static IP2NClient* createClient(const std::string& host, unsigned short port, int verbosity, const std::string& extras);
 
     /** Destroys a client.
     */
@@ -104,7 +104,7 @@ public:
     /** Return port used.
     @remarks An implementation must be supplied for this method.
     */
-    virtual int getPort() = 0;
+    virtual unsigned short getPort() = 0;
     /** Return verbosity used.
     @remarks An implementation must be supplied for this method.
     */
